+++ CHANGELOG.md
@@ -5,7 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.0.0] - 2019-11-18
+## 2.1.0 - 2019-11-22
+
+### Changed
+- The getId() method has been changed to id() in the BaseEntity interface, because it should not be considered a getter.
+
+### Removed
+- The method setId(id) has been removed from the BaseEntity interface.
+- The method getDefaultOrder() has been removed from the CrudService interface. Now, the default order is the physical order in the database.
+
+### Fixed
+- Fixed total element assignment (totalSize) of paged list conversion in class EntityListOutputConverter.
+
+## 2.0.0 - 2019-11-18
 
 ### Added
 - New list method overloads for full listing, with and without sorting.